requirements.txt
discord.py==2.5.2
aiohttp
python-dotenv
